transfer template session IT and fix some bug
