+++ confignode/src/main/java/org/apache/iotdb/confignode/persistence/schema/ClusterSchemaInfo.java
@@ -80,6 +80,9 @@ import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 
+import static org.apache.iotdb.commons.conf.IoTDBConstant.ONE_LEVEL_PATH_WILDCARD;
+import static org.apache.iotdb.db.metadata.MetadataConstant.ALL_TEMPLATE;
+
 /**
  * The ClusterSchemaInfo stores cluster schema. The cluster schema including: 1. StorageGroupSchema
  * 2. Template (Not implement yet)
@@ -592,7 +595,12 @@ public class ClusterSchemaInfo implements SnapshotProcessor {
     TemplateInfoResp result = new TemplateInfoResp();
     List<Template> list = new ArrayList<>();
     try {
-      list.add(templateTable.getTemplate(getSchemaTemplatePlan.getTemplateName()));
+      String templateName = getSchemaTemplatePlan.getTemplateName();
+      if (templateName.equals(ONE_LEVEL_PATH_WILDCARD)) {
+        list.addAll(templateTable.getAllTemplate());
+      } else {
+        list.add(templateTable.getTemplate(templateName));
+      }
       result.setTemplateList(list);
       result.setStatus(RpcUtils.getStatus(TSStatusCode.SUCCESS_STATUS));
     } catch (MetadataException e) {
@@ -652,7 +660,13 @@ public class ClusterSchemaInfo implements SnapshotProcessor {
     PathInfoResp pathInfoResp = new PathInfoResp();
     TSStatus status;
     try {
-      int templateId = templateTable.getTemplate(getPathsSetTemplatePlan.getName()).getId();
+      String templateName = getPathsSetTemplatePlan.getName();
+      int templateId;
+      if (templateName.equals(ONE_LEVEL_PATH_WILDCARD)) {
+        templateId = ALL_TEMPLATE;
+      } else {
+        templateId = templateTable.getTemplate(templateName).getId();
+      }
       pathInfoResp.setPathList(mTree.getPathsSetOnTemplate(templateId, false));
       status = new TSStatus(TSStatusCode.SUCCESS_STATUS.getStatusCode());
     } catch (MetadataException e) {
